Actualizar registro de personas

+++ src/main/java/com/mayte/Aprendiendo/interfaceService/IpersonaService.java
@@ -7,7 +7,7 @@ import com.mayte.Aprendiendo.models.person;
 
 public interface IpersonaService {
 	public List<person>listar();
-	public Optional<person>listarId(int id);
+	public Optional<person>listarId(Long id);
 	public int save(person p);
 	public void delete(int id);		
 }
